ci: introduce .test.mk instead of .travis.mk
This patch introduces the new test makefile `.test.mk` instead of the
ancient `.travis.mk`. The `.test.mk` file is a fully reworked version of
`.travis.mk` with the huge refactoring and contains only those targets
that are indeed used. But for now it doesn't have targets for the LuaJit
integration testing that are planned to add later.

NO_DOC=ci
NO_TEST=ci
NO_CHANGELOG=ci